requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
stripe==7.0.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
